ai/requirements.txt
numpy>=2.1.0
pymongo>=4.9.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.9.0
sentence-transformers>=3.2.0
pypdf>=4.3.1